Factorizing python code.
Explicit png names for Flame speed depending on:
   creation with Xdir or Ydir script.

Correct matplotlib warning by using handles for legend.

Update direction in figure(png) name.

Explicit png names for Flame speed.
